API change (float -> double) and inspections (cherry picked from commit 6970224)

+++ beam-avhrr-reader/src/main/java/org/esa/beam/dataio/avhrr/BandReader.java
@@ -28,7 +28,7 @@ public interface BandReader {
 
     public String getBandDescription();
 
-    public float getScalingFactor();
+    public double getScalingFactor();
 
     public int getDataType();
